third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-clip.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="target1" style="clip: rect(auto, auto, auto, auto);"></div>
<div id="target2" style="clip: rect(auto, auto, auto, 5);"></div>
<div id="target3" style="clip: rect(5, 2, auto, 5);"></div>
<script>
test(function() {
  // Tests https://code.google.com/p/chromium/issues/detail?id=412921
  assert_equals(getComputedStyle(target1).clip, 'rect(auto, auto, auto, auto)');
  assert_equals(getComputedStyle(target2).clip, 'rect(auto, auto, auto, 5px)');
  assert_equals(getComputedStyle(target3).clip, 'rect(5px, 2px, auto, 5px)');
});
</script>
